+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support old versions of react (pre v16.14)
 - Prefer idiomatic phrasing by default (use `numeric: 'always'` for the old behaviour)
+- Added a context provider to allow default options to be specified
 
 ## 2.0.0 (2024-02-09)